Only subscribe to updates for entities we're interested in
This uses a filter argument to the websocket subscription function to
only subscribe to entities that are currently being displayed on the
device screen.

The Home Assistant API for this is not currently implemented in the
client library so we needed to patch the library slightly.

For more details about this API see:
https://github.com/home-assistant/home-assistant-js-websocket/issues/511