Fix the use of custom display configuration files
Axios when running within Node.js doesn't support the file:// protocol.
This changes it so that it explicitly uses Node.js's `fs` package to
load local files if `file://` is detected in the URI.

This also fixes the fact that the custom display configuration was not
being loaded on save, only on startup of the plugin.

It also fixes it so that if you remove a custom display configuration it
will fall back to the default configuration without needing to restart
the plugin.

This fixes #342